+++ Frontend/public/static/templates/detector.html
