+++ Frontend/public/static/templates/index.html
@@ -16,7 +16,7 @@
         height:100vh;
         background-size:cover;
         position:relative;
-        background-image: url('../img/lung.jpeg');
+        background-image: url('./static/img/lung.jpeg');
     }
     /* text css above hero image*/
     .hero_title {
@@ -109,9 +109,6 @@
                         <a class="nav-link" href="#about">About Pneumonia</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="#impact">Impact</a>
-                    </li>
-                    <li class="nav-item">
                         <a class="nav-link" href="#example">Example</a>
                     </li>
                 </ul>
@@ -146,21 +143,6 @@
                 </div>
     </section>
 
-    <!-- impact section-->
-    <section id="impact">
-        <div class="container mt-4 pt-4">
-            <h1 class="text-center">Impact</h1>
-            <div class="row mt-4">
-                <div class="col-lg-12">
-                    <ul style="margin: 0px 40px;">
-                        <li>Pneumonia is often misdiagnosed—an NIH study found that nearly 28 out of 73 patients were misdiagnosed with pneumonia (38% inaccuracy).</li>
-                        <li>Misdiagnosis can cause health complications since patients are required to go through treatments that they would otherwise not need.</li>
-                        <li>Machine learning can help developing countries and people in rural areas where there may be a limited number of physicians available to diagnose.</li>
-                    </ul>
-                </div>
-    </section>
-
-
     <!-- example section-->
     <section id="example">
         <div class="container mt-4 pt-4">
@@ -168,7 +150,7 @@
             <div class="row mt-4">
                 <div class="col-lg-12">
                     <p style="text-align: center;">When choosing a file to submit, ensure that it is titled "LungXray.jpg". <em><strong>No other filetype will be accepted.</strong></em></p>
-                    <form action="detector.html" method="POST" style="margin: auto;">
+                    <form action="uploader" method="POST" style="margin: auto;">
                         <input type="file" id="myFile" name="filename" onchange="loadFile(event)" accept=".jpg"><br><br>
                         <input type="submit"><br>
                         <p><img id="output" width="200" /></p>
